pyproject.toml
[project]
name = "nl2gh"
version = "0.1.0"
description = "Natural language to GitHub API query CLI"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.52.0",
    "google-genai>=1.0.0",
    "httpx>=0.27.0",
    "pydantic>=2.9.0",
    "typer>=0.12.0",
    "rich>=13.0.0",
    "python-dotenv>=1.1.0",
    "tenacity>=9.0.0",
]

[project.scripts]
nl2gh = "nl2gh.cli:app"

[project.optional-dependencies]
eval = [
    "pandas>=2.2.0",
    "tabulate>=0.9.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/nl2gh"]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
]
